backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
accelerate==0.25.0
requests==2.31.0
